flow: support EDA (#622)
* flow:update

* flow:add event post_change

* flow: support EDA

* remove invalid script
