+++ middleware/flow/src/api/ci/flow_ci_inst_api.rs
@@ -84,24 +84,4 @@ impl FlowCiInstApi {
 
         TardisResp::ok(Void {})
     }
-
-    /// refresh var_collect / 刷新var_collect
-    #[oai(path = "/refresh_var_collect", method = "get")]
-    async fn refresh_var_collect(&self) -> TardisApiResult<Void> {
-        let mut funs = flow_constants::get_tardis_inst();
-        tokio::spawn(async move {
-            funs.begin().await.unwrap();
-            match FlowInstServ::refresh_var_collect(&funs).await {
-                Ok(_) => {
-                    log::trace!("[Flow.Inst] add log success")
-                }
-                Err(e) => {
-                    log::warn!("[Flow.Inst] failed to add log:{e}")
-                }
-            }
-            funs.commit().await.unwrap();
-        });
-
-        TardisResp::ok(Void {})
-    }
 }
