+++ middleware/flow/src/serv/flow_inst_serv.rs
@@ -16,19 +16,17 @@ use bios_basic::{
     },
 };
 use itertools::Itertools;
-use serde_json::json;
 use tardis::{
     basic::{dto::TardisContext, result::TardisResult},
-    chrono::{DateTime, SecondsFormat, Utc},
+    chrono::{DateTime, Utc},
     db::sea_orm::{
         self,
         sea_query::{Alias, Cond, Expr, Query},
         JoinType, Set,
     },
     futures_util::future::join_all,
-    log::{debug, warn},
+    log::debug,
     serde_json::Value,
-    tokio,
     web::web_resp::TardisPage,
     TardisFuns, TardisFunsInst,
 };
@@ -43,18 +41,20 @@ use crate::{
             FlowInstTransitionInfo, FlowOperationContext,
         },
         flow_model_dto::{FlowModelDetailResp, FlowModelFilterReq},
-        flow_state_dto::{FlowStateDetailResp, FlowStateFilterReq, FlowStateRelModelExt, FlowSysStateKind},
+        flow_state_dto::{FlowStateFilterReq, FlowStateRelModelExt, FlowSysStateKind},
         flow_transition_dto::{
-            FlowTransitionActionByStateChangeInfo, FlowTransitionActionByVarChangeInfoChangedKind, FlowTransitionActionChangeAgg, FlowTransitionActionChangeInfo,
-            FlowTransitionActionChangeKind, FlowTransitionDetailResp, FlowTransitionFrontActionInfo, FlowTransitionFrontActionRightValue, StateChangeConditionOp, TagRelKind,
+            FlowTransitionDetailResp, FlowTransitionFrontActionInfo
         },
-        flow_var_dto::{FillType, FlowVarInfo},
+        flow_var_dto::FillType,
     },
-    flow_constants,
+    flow_config::FlowConfig,
+    flow_initializer::{default_flow_avatar, ws_flow_client},
     serv::{flow_model_serv::FlowModelServ, flow_state_serv::FlowStateServ},
 };
 
 use super::{
+    clients::event_client::FlowEventExt,
+    flow_event_serv::FlowEventServ,
     flow_external_serv::FlowExternalServ,
     flow_rel_serv::{FlowRelKind, FlowRelServ},
 };
@@ -264,33 +264,19 @@ impl FlowInstServ {
     }
 
     pub async fn modify_assigned(flow_inst_id: &str, assigned_id: &str, funs: &TardisFunsInst, ctx: &TardisContext) -> TardisResult<()> {
-        if funs
-            .db()
-            .count(
-                Query::select()
-                    .column((flow_inst::Entity, flow_inst::Column::Id))
-                    .from(flow_inst::Entity)
-                    .and_where(Expr::col((flow_inst::Entity, flow_inst::Column::Id)).eq(flow_inst_id.to_string()))
-                    .and_where(Expr::col((flow_inst::Entity, flow_inst::Column::OwnPaths)).like(format!("{}%", ctx.own_paths))),
-            )
-            .await?
-            == 0
-        {
-            return Err(funs.err().not_found(
-                "flow_inst",
-                "modify_assigned",
-                &format!("flow instance {} not found", flow_inst_id),
-                "404-flow-inst-not-found",
-            ));
+        if let Some(ws_client) = ws_flow_client().await {
+            ws_client
+                .publish_modify_assigned(
+                    flow_inst_id.to_string(),
+                    assigned_id.to_string(),
+                    default_flow_avatar().await.clone(),
+                    funs.conf::<FlowConfig>().invoke.spi_app_id.clone(),
+                    ctx,
+                )
+                .await
+        } else {
+            FlowEventServ::do_modify_assigned(flow_inst_id, assigned_id, ctx, funs).await
         }
-        let flow_inst = flow_inst::ActiveModel {
-            id: Set(flow_inst_id.to_string()),
-            current_assigned: Set(Some(assigned_id.to_string())),
-            ..Default::default()
-        };
-        funs.db().update_one(flow_inst, ctx).await?;
-
-        Ok(())
     }
 
     pub async fn get(flow_inst_id: &str, funs: &TardisFunsInst, ctx: &TardisContext) -> TardisResult<FlowInstDetailResp> {
@@ -655,22 +641,13 @@ impl FlowInstServ {
         funs: &TardisFunsInst,
         ctx: &TardisContext,
     ) -> TardisResult<FlowInstTransferResp> {
-        // record updated instance id
-        let mut updated_instance_list: Vec<String> = Vec::new();
-        let result = Self::do_transfer(flow_inst_id, transfer_req, &mut updated_instance_list, skip_filter, callback_kind, funs, ctx).await;
-
-        for updated_instance_id in updated_instance_list {
-            Self::do_front_change(&updated_instance_id, ctx, funs).await?;
-        }
-
-        result
+        Self::do_transfer(flow_inst_id, transfer_req, skip_filter, callback_kind, funs, ctx).await
     }
 
     #[async_recursion]
     async fn do_transfer(
         flow_inst_id: &str,
         transfer_req: &FlowInstTransferReq,
-        updated_instance_list: &mut Vec<String>,
         skip_filter: bool,
         callback_kind: FlowExternalCallbackOp,
         funs: &TardisFunsInst,
@@ -819,7 +796,6 @@ impl FlowInstServ {
         }
 
         funs.db().update_one(flow_inst, ctx).await?;
-        updated_instance_list.push(flow_inst_id.to_string());
 
         // get updated instance detail
         let flow_inst_detail = Self::get(flow_inst_id, funs, ctx).await?;
@@ -852,18 +828,42 @@ impl FlowInstServ {
         let post_changes =
             model_transition.into_iter().find(|model_transition| model_transition.id == next_flow_transition.next_flow_transition_id).unwrap_or_default().action_by_post_changes();
         if !post_changes.is_empty() {
-            Self::do_post_change(
-                &flow_inst_detail,
-                &flow_model,
-                &next_transition_detail,
-                &prev_flow_state,
-                &next_flow_state,
-                post_changes,
-                updated_instance_list,
-                ctx,
-                funs,
-            )
-            .await?;
+            if let Some(ws_client) = ws_flow_client().await {
+                ws_client
+                    .publish_post_change(
+                        flow_inst_detail.id.clone(),
+                        next_transition_detail.id.clone(),
+                        default_flow_avatar().await.clone(),
+                        funs.conf::<FlowConfig>().invoke.spi_app_id.clone(),
+                        ctx,
+                    )
+                    .await?;
+            } else {
+                FlowEventServ::do_post_change(&flow_inst_detail.id, &next_transition_detail.id, ctx, funs).await?;
+                // Self::do_post_change(
+                //     &flow_inst_detail,
+                //     &flow_model,
+                //     &next_transition_detail,
+                //     &prev_flow_state,
+                //     &next_flow_state,
+                //     post_changes,
+                //     ctx,
+                //     funs,
+                // )
+                // .await?;
+            }
+        }
+        if let Some(ws_client) = ws_flow_client().await {
+            ws_client
+                .publish_front_change(
+                    flow_inst_id.to_string(),
+                    default_flow_avatar().await.clone(),
+                    funs.conf::<FlowConfig>().invoke.spi_app_id.clone(),
+                    ctx,
+                )
+                .await?;
+        } else {
+            FlowEventServ::do_front_change(flow_inst_id, ctx, funs).await?;
         }
 
         Self::gen_transfer_resp(flow_inst_id, &prev_flow_state.id, ctx, funs).await
@@ -940,265 +940,6 @@ impl FlowInstServ {
         })
     }
 
-    /// handling post change when the transition occurs
-    async fn do_post_change(
-        current_inst: &FlowInstDetailResp,
-        current_model: &FlowModelDetailResp,
-        transition_detail: &FlowTransitionDetailResp,
-        prev_flow_state: &FlowStateDetailResp,
-        next_flow_state: &FlowStateDetailResp,
-        post_changes: Vec<FlowTransitionActionChangeInfo>,
-        updated_instance_list: &mut Vec<String>,
-        ctx: &TardisContext,
-        funs: &TardisFunsInst,
-    ) -> TardisResult<()> {
-        for post_change in post_changes {
-            let post_change = FlowTransitionActionChangeAgg::from(post_change);
-            match post_change.kind {
-                FlowTransitionActionChangeKind::Var => {
-                    if let Some(mut change_info) = post_change.var_change_info {
-                        if change_info.changed_kind.is_some() && change_info.changed_kind.clone().unwrap() == FlowTransitionActionByVarChangeInfoChangedKind::AutoGetOperateTime {
-                            change_info.changed_val = Some(json!(Utc::now().to_rfc3339_opts(SecondsFormat::Millis, true)));
-                            change_info.changed_kind = Some(FlowTransitionActionByVarChangeInfoChangedKind::ChangeContent);
-                        }
-                        let rel_tag = change_info.obj_tag.unwrap_or_default();
-                        if !rel_tag.is_empty() {
-                            let mut resp = FlowExternalServ::do_fetch_rel_obj(
-                                &current_model.tag,
-                                &current_inst.id,
-                                &current_inst.rel_business_obj_id,
-                                vec![(rel_tag.clone(), change_info.obj_tag_rel_kind.clone())],
-                                ctx,
-                                funs,
-                            )
-                            .await?;
-                            if !resp.rel_bus_objs.is_empty() {
-                                for rel_bus_obj_id in resp.rel_bus_objs.pop().unwrap().rel_bus_obj_ids {
-                                    let inst_id = Self::get_inst_ids_by_rel_business_obj_id(vec![rel_bus_obj_id.clone()], funs, ctx).await?.pop().unwrap_or_default();
-                                    FlowExternalServ::do_modify_field(
-                                        &rel_tag,
-                                        transition_detail,
-                                        &rel_bus_obj_id,
-                                        &inst_id,
-                                        FlowExternalCallbackOp::PostAction,
-                                        next_flow_state.name.clone(),
-                                        next_flow_state.sys_state.clone(),
-                                        prev_flow_state.name.clone(),
-                                        prev_flow_state.sys_state.clone(),
-                                        vec![FlowExternalParams {
-                                            rel_kind: None,
-                                            rel_tag: None,
-                                            var_id: None,
-                                            var_name: Some(change_info.var_name.clone()),
-                                            value: change_info.changed_val.clone(),
-                                            changed_kind: change_info.changed_kind.clone(),
-                                        }],
-                                        ctx,
-                                        funs,
-                                    )
-                                    .await?;
-                                    updated_instance_list.push(inst_id.clone());
-                                }
-                            }
-                        } else {
-                            FlowExternalServ::do_modify_field(
-                                &current_model.tag,
-                                transition_detail,
-                                &current_inst.rel_business_obj_id,
-                                &current_inst.id,
-                                FlowExternalCallbackOp::PostAction,
-                                next_flow_state.name.clone(),
-                                next_flow_state.sys_state.clone(),
-                                prev_flow_state.name.clone(),
-                                prev_flow_state.sys_state.clone(),
-                                vec![FlowExternalParams {
-                                    rel_kind: None,
-                                    rel_tag: None,
-                                    var_id: None,
-                                    var_name: Some(change_info.var_name.clone()),
-                                    value: change_info.changed_val.clone(),
-                                    changed_kind: change_info.changed_kind,
-                                }],
-                                ctx,
-                                funs,
-                            )
-                            .await?;
-                            updated_instance_list.push(current_inst.id.clone());
-                        }
-                    }
-                }
-                FlowTransitionActionChangeKind::State => {
-                    if let Some(change_info) = post_change.state_change_info {
-                        let mut resp = FlowExternalServ::do_fetch_rel_obj(
-                            &current_model.tag,
-                            &current_inst.id,
-                            &current_inst.rel_business_obj_id,
-                            vec![(change_info.obj_tag.clone(), change_info.obj_tag_rel_kind.clone())],
-                            ctx,
-                            funs,
-                        )
-                        .await?;
-                        if !resp.rel_bus_objs.is_empty() {
-                            let inst_ids = Self::find_inst_ids_by_rel_obj_ids(current_model, resp.rel_bus_objs.pop().unwrap().rel_bus_obj_ids, &change_info, funs, ctx).await?;
-                            Self::do_modify_state_by_post_action(inst_ids, &change_info, updated_instance_list, funs, ctx).await?;
-                        }
-                    }
-                }
-            }
-        }
-
-        Ok(())
-    }
-    async fn find_inst_ids_by_rel_obj_ids(
-        flow_model: &FlowModelDetailResp,
-        rel_bus_obj_ids: Vec<String>,
-        change_info: &FlowTransitionActionByStateChangeInfo,
-        funs: &TardisFunsInst,
-        ctx: &TardisContext,
-    ) -> TardisResult<Vec<String>> {
-        let mut result_rel_obj_ids = Self::filter_rel_obj_ids_by_state(&rel_bus_obj_ids, &change_info.obj_current_state_id, funs, ctx).await?;
-
-        if let Some(change_condition) = change_info.change_condition.clone() {
-            // Check mismatch rel_obj_ids and filter them
-            let mut mismatch_rel_obj_ids = vec![];
-            for rel_obj_id in result_rel_obj_ids.iter() {
-                if change_condition.current {
-                    // collect rel tags
-                    let mut rel_tags = vec![];
-                    for condition_item in change_condition.conditions.iter() {
-                        if condition_item.obj_tag.is_some() && !condition_item.state_id.is_empty() {
-                            rel_tags.push((condition_item.obj_tag.clone().unwrap(), condition_item.obj_tag_rel_kind.clone()));
-                        }
-                    }
-                    let inst_id = Self::get_inst_ids_by_rel_business_obj_id(vec![rel_obj_id.clone()], funs, ctx).await?.pop().unwrap_or_default();
-                    let tag = if change_info.obj_tag_rel_kind == Some(TagRelKind::ParentOrSub) {
-                        &flow_model.tag
-                    } else {
-                        &change_info.obj_tag
-                    };
-                    let resp = FlowExternalServ::do_fetch_rel_obj(tag, &inst_id, rel_obj_id, rel_tags, ctx, funs).await?;
-                    if !resp.rel_bus_objs.is_empty() {
-                        for rel_bus_obj in resp.rel_bus_objs {
-                            let condition = change_condition
-                                .conditions
-                                .iter()
-                                .find(|condition| condition.obj_tag.is_some() && condition.obj_tag.clone().unwrap() == rel_bus_obj.rel_tag.clone())
-                                .unwrap();
-                            let rel_obj_ids = Self::filter_rel_obj_ids_by_state(&rel_bus_obj.rel_bus_obj_ids, &Some(condition.state_id.clone()), funs, ctx).await?;
-                            match condition.op {
-                                StateChangeConditionOp::And => {
-                                    if rel_bus_obj.rel_bus_obj_ids.is_empty() || rel_bus_obj.rel_bus_obj_ids.len() != rel_obj_ids.len() {
-                                        mismatch_rel_obj_ids.push(rel_obj_id.clone());
-                                        continue;
-                                    }
-                                }
-                                StateChangeConditionOp::Or => {
-                                    if rel_obj_ids.is_empty() {
-                                        mismatch_rel_obj_ids.push(rel_obj_id.clone());
-                                        continue;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            result_rel_obj_ids = result_rel_obj_ids.into_iter().filter(|result_rel_obj_id| !mismatch_rel_obj_ids.contains(result_rel_obj_id)).collect_vec();
-        }
-
-        let result = Self::get_inst_ids_by_rel_business_obj_id(result_rel_obj_ids, funs, ctx).await?;
-        Ok(result)
-    }
-
-    async fn filter_rel_obj_ids_by_state(
-        rel_bus_obj_ids: &Vec<String>,
-        obj_current_state_id: &Option<Vec<String>>,
-        funs: &TardisFunsInst,
-        _ctx: &TardisContext,
-    ) -> TardisResult<Vec<String>> {
-        #[derive(sea_orm::FromQueryResult)]
-        pub struct FlowInstRelObjIdsResult {
-            pub id: String,
-            pub current_state_id: String,
-            pub rel_business_obj_id: String,
-        }
-        let rel_insts = funs
-            .db()
-            .find_dtos::<FlowInstRelObjIdsResult>(
-                Query::select()
-                    .columns([flow_inst::Column::Id, flow_inst::Column::CurrentStateId, flow_inst::Column::RelBusinessObjId])
-                    .from(flow_inst::Entity)
-                    .and_where(Expr::col(flow_inst::Column::RelBusinessObjId).is_in(rel_bus_obj_ids.clone())),
-            )
-            .await?;
-        if rel_bus_obj_ids.len() != rel_insts.len() {
-            return Err(funs.err().not_found("flow_inst", "do_post_change", "some flow instances not found", "404-flow-inst-not-found"));
-        }
-        Ok(rel_insts
-            .iter()
-            .filter(|inst_result| {
-                if let Some(obj_current_state_id) = obj_current_state_id.clone() {
-                    if !obj_current_state_id.is_empty() && !obj_current_state_id.contains(&inst_result.current_state_id) {
-                        return false;
-                    }
-                }
-                true
-            })
-            .map(|inst_result| inst_result.rel_business_obj_id.clone())
-            .collect_vec())
-    }
-
-    async fn do_modify_state_by_post_action(
-        rel_inst_ids: Vec<String>,
-        change_info: &FlowTransitionActionByStateChangeInfo,
-        updated_instance_list: &mut Vec<String>,
-        funs: &TardisFunsInst,
-        ctx: &TardisContext,
-    ) -> TardisResult<()> {
-        let insts = Self::find_detail(rel_inst_ids, funs, ctx).await?;
-        for rel_inst in insts {
-            // find transition
-            let flow_model = FlowModelServ::get_item(
-                &rel_inst.rel_flow_model_id,
-                &FlowModelFilterReq {
-                    basic: RbumBasicFilterReq {
-                        with_sub_own_paths: true,
-                        own_paths: Some("".to_string()),
-                        ..Default::default()
-                    },
-                    ..Default::default()
-                },
-                funs,
-                ctx,
-            )
-            .await?;
-            let transition_resp = Self::do_find_next_transitions(&rel_inst, &flow_model, None, &None, false, funs, ctx)
-                .await?
-                .next_flow_transitions
-                .into_iter()
-                .filter(|transition_detail| *transition_detail.next_flow_state_id == change_info.changed_state_id)
-                .collect_vec()
-                .pop();
-            if let Some(transition) = transition_resp {
-                Self::do_transfer(
-                    &rel_inst.id,
-                    &FlowInstTransferReq {
-                        flow_transition_id: transition.next_flow_transition_id,
-                        message: None,
-                        vars: None,
-                    },
-                    updated_instance_list,
-                    true,
-                    FlowExternalCallbackOp::PostAction,
-                    funs,
-                    ctx,
-                )
-                .await?;
-            }
-        }
-        Ok(())
-    }
-
     /// request webhook when the transition occurs
     async fn do_request_webhook(from_transition_detail: Option<&FlowTransitionDetailResp>, to_transition_detail: Option<&FlowTransitionDetailResp>) -> TardisResult<()> {
         if let Some(from_transition_detail) = from_transition_detail {
@@ -1226,7 +967,7 @@ impl FlowInstServ {
     }
 
     /// The kernel function of flow processing
-    async fn do_find_next_transitions(
+    pub async fn do_find_next_transitions(
         flow_inst: &FlowInstDetailResp,
         flow_model: &FlowModelDetailResp,
         spec_flow_transition_id: Option<String>,
@@ -1360,129 +1101,22 @@ impl FlowInstServ {
             current_vars: Set(Some(TardisFuns::json.obj_to_json(&new_vars)?)),
             ..Default::default()
         };
+        funs.db().update_one(flow_inst, ctx).await.unwrap();
 
-        let flow_inst_id_sync = flow_inst_id.to_string();
-        let ctx_sync = ctx.clone();
-        tokio::spawn(async move {
-            let mut funs = flow_constants::get_tardis_inst();
-            funs.db().update_one(flow_inst, &ctx_sync).await.unwrap();
-            funs.begin().await.unwrap();
-            match Self::do_front_change(&flow_inst_id_sync, &ctx_sync, &funs).await {
-                Ok(_) => {}
-                Err(e) => {
-                    warn!("[Flow.Inst] failed to add log:{e}")
-                }
-            }
-            funs.commit().await.unwrap();
-        });
-
-        Ok(())
-    }
-
-    #[async_recursion]
-    async fn do_front_change(flow_inst_id: &str, ctx: &TardisContext, funs: &TardisFunsInst) -> TardisResult<()> {
-        let flow_inst_detail = Self::get(flow_inst_id, funs, ctx).await?;
-        let flow_model = FlowModelServ::get_item(
-            &flow_inst_detail.rel_flow_model_id,
-            &FlowModelFilterReq {
-                basic: RbumBasicFilterReq {
-                    with_sub_own_paths: true,
-                    own_paths: Some("".to_string()),
-                    ..Default::default()
-                },
-                ..Default::default()
-            },
-            funs,
-            ctx,
-        )
-        .await?;
-        let flow_transitions = flow_model
-            .transitions()
-            .into_iter()
-            .filter(|trans| trans.from_flow_state_id == flow_inst_detail.current_state_id && !trans.action_by_front_changes().is_empty())
-            .sorted_by_key(|trans| trans.sort)
-            .collect_vec();
-        if flow_transitions.is_empty() {
-            return Ok(());
-        }
-        for flow_transition in flow_transitions {
-            if Self::check_front_conditions(&flow_inst_detail, flow_transition.action_by_front_changes())? {
-                Self::transfer(
-                    &flow_inst_detail.id,
-                    &FlowInstTransferReq {
-                        flow_transition_id: flow_transition.id.clone(),
-                        message: None,
-                        vars: None,
-                    },
-                    true,
-                    FlowExternalCallbackOp::ConditionalTrigger,
-                    funs,
+        if let Some(ws_client) = ws_flow_client().await {
+            ws_client
+                .publish_front_change(
+                    flow_inst_id.to_string(),
+                    default_flow_avatar().await.clone(),
+                    funs.conf::<FlowConfig>().invoke.spi_app_id.clone(),
                     ctx,
                 )
                 .await?;
-                break;
-            }
+        } else {
+            FlowEventServ::do_front_change(flow_inst_id, ctx, funs).await?;
         }
 
         Ok(())
-    }
-
-    fn check_front_conditions(flow_inst_detail: &FlowInstDetailResp, conditions: Vec<FlowTransitionFrontActionInfo>) -> TardisResult<bool> {
-        if flow_inst_detail.current_vars.is_none() {
-            return Ok(false);
-        }
-        let current_vars = flow_inst_detail.current_vars.clone().unwrap();
-        for condition in conditions {
-            if !Self::do_check_front_condition(&current_vars, &condition)? {
-                return Ok(false);
-            }
-        }
-
-        Ok(true)
-    }
-
-    fn do_check_front_condition(current_vars: &HashMap<String, Value>, condition: &FlowTransitionFrontActionInfo) -> TardisResult<bool> {
-        match condition.right_value {
-            FlowTransitionFrontActionRightValue::ChangeContent => {
-                if let Some(left_value) = current_vars.get(&condition.left_value) {
-                    Ok(condition.relevance_relation.check_conform(
-                        left_value.as_str().unwrap_or(left_value.to_string().as_str()).to_string(),
-                        condition
-                            .change_content
-                            .clone()
-                            .unwrap_or_default()
-                            .as_str()
-                            .unwrap_or(condition.change_content.clone().unwrap_or_default().to_string().as_str())
-                            .to_string(),
-                    ))
-                } else {
-                    Ok(false)
-                }
-            }
-            FlowTransitionFrontActionRightValue::SelectField => {
-                if let (Some(left_value), Some(right_value)) = (
-                    current_vars.get(&condition.left_value),
-                    current_vars.get(&condition.select_field.clone().unwrap_or_default()),
-                ) {
-                    Ok(condition.relevance_relation.check_conform(
-                        left_value.as_str().unwrap_or(left_value.to_string().as_str()).to_string(),
-                        right_value.as_str().unwrap_or(left_value.to_string().as_str()).to_string(),
-                    ))
-                } else {
-                    Ok(false)
-                }
-            }
-            FlowTransitionFrontActionRightValue::RealTime => {
-                if let Some(left_value) = current_vars.get(&condition.left_value) {
-                    Ok(condition.relevance_relation.check_conform(
-                        left_value.as_str().unwrap_or(left_value.to_string().as_str()).to_string(),
-                        Utc::now().to_rfc3339_opts(SecondsFormat::Millis, true),
-                    ))
-                } else {
-                    Ok(false)
-                }
-            }
-        }
     }
 
     async fn get_new_vars(flow_inst_id: &str, funs: &TardisFunsInst, ctx: &TardisContext) -> TardisResult<Value> {
@@ -1570,59 +1204,6 @@ impl FlowInstServ {
                     &ctx,
                 )
                 .await?;
-            }
-        }
-
-        Ok(())
-    }
-
-    pub async fn refresh_var_collect(funs: &TardisFunsInst) -> TardisResult<()> {
-        #[derive(sea_orm::FromQueryResult)]
-        pub struct FlowTransitionsResult {
-            id: String,
-            vars_collect: Value,
-        }
-
-        //{"filed_type":"built","value_type":"tag","global":"true","tag_name":"priority","digits":null}
-        #[derive(serde::Serialize, serde::Deserialize, Debug, Default, Clone)]
-        pub struct FlowTransitionsExt {
-            filed_type: String,
-            value_type: String,
-            tag_name: Option<String>,
-            tag: Option<String>,
-            global: Option<String>,
-            digits: Option<String>,
-        }
-        let global_ctx = TardisContext::default();
-        let flow_transition_list = funs
-            .db()
-            .find_dtos::<FlowTransitionsResult>(Query::select().columns([flow_transition::Column::Id, flow_transition::Column::VarsCollect]).from(flow_transition::Entity))
-            .await?
-            .into_iter()
-            .filter(|res| !TardisFuns::json.json_to_obj::<Vec<FlowVarInfo>>(res.vars_collect.clone()).unwrap_or_default().is_empty())
-            .collect_vec();
-        for flow_transition in flow_transition_list {
-            let mut is_update = false;
-            let mut vars_collect = TardisFuns::json.json_to_obj::<Vec<FlowVarInfo>>(flow_transition.vars_collect.clone())?;
-            for var in vars_collect.iter_mut() {
-                let mut ext = TardisFuns::json.str_to_obj::<FlowTransitionsExt>(&var.ext.clone().unwrap_or_default()).unwrap_or_default();
-                if ext.tag_name.is_some() {
-                    ext.tag_name = Some(format!("_:{}", ext.tag_name.unwrap_or_default()));
-                    var.ext = Some(TardisFuns::json.obj_to_string(&ext)?);
-                    is_update = true;
-                }
-            }
-            if is_update {
-                funs.db()
-                    .update_one(
-                        flow_transition::ActiveModel {
-                            id: Set(flow_transition.id.clone()),
-                            vars_collect: Set(TardisFuns::json.obj_to_json(&vars_collect)?),
-                            ..Default::default()
-                        },
-                        &global_ctx,
-                    )
-                    .await?;
             }
         }
