+++ support/iam/src/console_tenant/api/iam_ct_app_set_api.rs
@@ -210,14 +210,4 @@ impl IamCtAppSetApi {
         ctx.execute_task().await?;
         TardisResp::ok(result)
     }
-
-    /// refresh data
-    #[oai(path = "/refresh_data", method = "get")]
-    async fn refresh_data(&self, ctx: TardisContextExtractor, request: &Request) -> TardisApiResult<Void> {
-        add_remote_ip(request, &ctx.0).await?;
-        let funs = iam_constants::get_tardis_inst();
-        IamSetServ::refresh_app_groups_data(&funs, &ctx.0).await?;
-        ctx.0.execute_task().await?;
-        TardisResp::ok(Void {})
-    }
 }
